Add functions for handling jsonb, and regression tests.
The functions and operators are identical to the json operators
and functions, except that the generator functions are not present.

Mostly they work by turning the jsonb back into json and then using the
same processing, so almost all the code is in jsonfuncs.c. This isn't
the way we want to go in the long term, but it gives us a stake in the
ground, so we can now go back and one by one reimplement these functions
using the jsonb API.