Fix contains operator in jsonb GIN support routine
Have gin_extract_jsonb() flag boths keys and elements as keys.  I
believe this is safe, but additional review would be useful.

Add some regression testing for this.  The same bug exists for GiST, and
that isn't fixed here, so the failing regression test adding for the
analogous GiST bug earlier continues to fail.