Introduce "superheader" abstraction
This makes it clearer how the nested structure can be interacted with
through either a varlena Jsonb buffer pointer, or a (nested) pointer to
a Jentry (which has a header...so we might be accessing that through
what is nominally a superheader pointer, and it doesn't matter).

This makes the structure of much of jsonb_support.c a lot clearer.