Clean-up "compression walking"
This doesn't have much to do with actual compression.  It actually
concerns converting an arbitrary JsonbValue to a 4 byte uncompressed
varlena representation (i.e. a Jsonb).  So stop using the misleading
term  compression, and start calling it conversion.  Use the type system
sensibly.  Clean-up and comment the code more generally.